requirements.txt
flask==2.3.3
pytube==15.0.0
moviepy==1.0.3
numpy==1.24.3
transformers==4.36.0
torch==2.0.1
python-dotenv==1.0.0